# documentation/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "homework-9",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "variables": {}
    }
  ],
  "paths": {
...
    "/movie": {
      "post": {
        "tags": [
          "Movies"
        ],
        "summary": "create movie",
        "operationId": "createmovie",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MTA0LCJlbWFpbCI6ImFsdmluIiwiZ2VuZGVyIjoibGFraSIsInBhc3N3b3JkIjoiMTIzIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNjk2Njc1MjgzfQ.yadfhGC04m8_7oaXZ8AO0RC5C4yPvAPxs4YEBwVRBgM"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/createmovierequest"
                  },
                  {
                    "example": {
                      "title": "testing title",
                      "genres": "testing genres",
                      "year": 2000
                    }
                  }
                ]
              },
              "example": {
                "title": "testing title",
                "genres": "testing genres",
                "year": 2000
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": []
      },
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "get movies with limit",
        "operationId": "getmovieswithlimit",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 5
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "enum": [
                "application/json"
              ],
              "type": "string",
              "example": "application/json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "createmovierequest": {
        "title": "createmovierequest",
        "required": [
          "title",
          "genres",
          "year"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "genres": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          }
        },
        "example": {
          "title": "testing title",
          "genres": "testing genres",
          "year": 2000
        }
      },
...
    }
  },
...
}